:boom: Refine `GinBrowse` command
`++repository` option is added to control the behavior of the command.
If this option is given, the command opens the repository page instead
of the blob or commit page.

Additionally, `--path` option is removed from the command. The command
now uses the second argument as the path to open the page. If the
argument is not given, the command uses the current buffer or the
current working directory if the current buffer is not a file.